macros/package: Use `?include=` for crates.io requests
By default that crates.io endpoint is including a bunch of related information that we don't actually need for this template. This extra information is slowing down the responses since additional database queries are required. With this commit the server regularly responds in 1-2ms while previously it was 5-10ms. This brought down the build time from 100s to 35s for me.